Move InfoButton to common
